Removed implemented (not yet tested)
It does not merge underfull leaves or internal nodes like it should but
it will removed keys from the tree. Honestly, I am unsure of the overall
utility (with respect to the complexity) of the underfull merge. Whether
or not it will be helpful will depend on the workload. I think I will
leave it out for the moment.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>