Stress tested pure block split, fixed bug.
The bug was subtle and may also be in the B+Tree in the data-structures
repo. It occurs when you have a ton of pure block splits. There were
actually two bugs, one where the wrong block was returned a "b" the
other where the newly created block was not inserted into the correct
location in the linked list.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>