Removed data-structures
There was no cross functional code that could be imported unfortunately.
The only way to do that would be to create a facade in front for the
BpTree API which is close but quite different than the in memory
version. For one thing, most API methods return an error. Why? because
the fmap subsystem could fail to load a block. Additionally keys and
value are []byte in this repository but in data-structures they are
arbitrary interfaces. At the moment it is better to just keep the
completely separated.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>