Benchmark and performance improvements
I wrote a benchmark (bench_test.go) and did some performance work based
on its findings.

1. Removed Crc32 checksum from fmap headers. This was a good debugging
tool but it takes a long time to do on every block load. Easy win.

2. Added a pool for [][]byte allocations in attachLeaf and
attachInternal. This was a big win.

3. Changed how the attachent work to reduce unneeded allocations.
Another win.

Overall, these changes sped up the system by an order of magnatude.
However, it is still way slower than I want (~20,000,000 ns/op). The big
culprit is reattachLeaf. Not sure if there is a way around that.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>