Node object chaching
From profiling it was discovered that attaching (and reattaching) leaves
and internal nodes was a big bottle neck. A lot of work was done
optimizing it but there were limits on what could be done.

To solve the problem now there is cache for the node objects. It is
important to clarify something about how this cache works. The
relationship between the objects and the backing store looks like this:

  ptr to node -> [ (leaf|internal) struct ] => [ mmap file ]

In other words the node structs just point into the mmap file. They are
not stored there themselves.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>