Another fix for almost-pure bug
This is getting complex. I don't have a test to replicate the bug but
here is a description:

get a block pretty full with a mixture of a large run of pure with a run
and some non-pure. Make it so that a balance will result in one smaller
item on one side and the rest on the other.

Then when an insert happens the key should be somewhere in the middle of
the almost full side. causing an out of order insert. This commit should
fix that.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>