Refactored packaged structure
The varchar package could be generally useful outside of the b+ tree so
I moved it out of the bptree packaged. I also had to expose the
constants so they could be used in both packages. This will also be
useful as I add more structures to this repository. If you have multiple
structures in one file they should hopefully be able to coexist since
their constants will not clash.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>