First cut and free for varchar
The first cut at the allocator for varchar will be very simple. It will
just be a coallescing free list sorted on the position of the blocks.
This will be simple to implement but has numerous run time
disadvantages. Namely, both free and alloc will be O(N) of the list. We
can improve that to O(lg(N)) for both by using two AVL trees. One will
be used to find nodes during alloc (by size) the other will be used to
find nodes for free (and coallesce) this will be keyed by address.

There may be better ways to arrange that. I should really read the
literature.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>