Defined a strategy for varchar key management
Going down it is the actual key. Once it gets to the leaf it is
converted to a pointer. The pointer then gets sent back up. Internal
nodes only work with the pointer never with the actual key. That means
when they want to do comparisons they need to indirect into the varchar
store.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>