Implemented New, Append, Get and Set for mmlist
This shows the main structure how this list is going to work. There is a
B+Tree which indexes to blks which hold the list. In theory it does not
have to be a B+Tree which goes down to the list index block. For
instance this would be a good opportunity to implement a bottom up
system like index sequential files. However, the B+Tree is already there
and it should have reasonable performance for this use case.

The actual data is stored in a varchar. So there are lots of B+Trees
here!

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>