This commit turns off some run time checking
The ordering of the blocks will no longer be checked after each block
update. This means that

1. More performance
2. More danger.

The checks that are disabled could be done on demand by running
Verify() which performs them as well. However, you will not catch the
errors at creation time. Hopefully, there are no errors to catch! I am
going to figure out how to make the whole thing have a debugging mode.
That way stuff doesn't have to be commented out.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>