Fix for a very rare (potential) insert bug
I haven't actually seen an instance of this bug but it could happen.
There is a whole other option that involves having remove consistently
merge blocks together. However, this is actually pretty complex in the
face of duplicate keys (which is why merging blocks was not
implemented). The duplicate key solution creates a lot of complexity
around when you can and cannot merge blocks.

Furthermore, trying to do clever things like: "pure block runs must
always be compact", means higher insert costs and higher remove costs.
As does: "values of duplicate keys should always be in sort order". As
must as I would like to see that as a feature, it will cause degraded
insert and remove performance.

A better idea would be to have subtrees for sorted values as a possible
mode. I don't know how I feel about that yet.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>