fixed bug in bptree/get.go:forward()
The bug goes like this. If you are looking in a range [5,7] but the tree
ends with key 4, getStart() which must always return something will
return the location of the last key 4. The correct behavior is to return
an empty bpt_iterator however what actually happens it is it will return
an iterator which will return just that last key and the terminate.

This causes things like Find() to erroneously return data if you search
for a range which is past the end of the tree.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>