Add use-parameterized-serialization to fs2-generic
The idea here is that sometimes serializers have to be specialized to
each instance of the tree. This option let's you do that by pushing them
into the struct rather than being static. The types are still
specialized to each tree but the serializers for those types are
supplied at instantiation of the tree object.

Signed-off-by: Tim Henderson <tim.tadh@gmail.com>